Added oneshot pack*() and unpack*() variants of all functions
these are useful for one-off encoding of simple stuff.
While it's possible to write a utility method that accepts a first-class
callable and creates a throwaway buffer internally, like

function encodeOneOff(T $value, Closure $writeFunc) : string{
	$writer = new ByteBufferWriter();
	$writeFunc($writer, $value);
	return $writer->getBuffer();
}

this is quite a bit slower than even unpack() for simple types, as well
as being mildly inconvenient.
So having dedicated one-off functions like in `Binary::` that aren't
slow is a good idea.