Avoid dynamic memcpy in varint encoding
This improves the write performance by about 12% in local tests, putting
varints almost on par with fixed integers (at least from PHP land).

In order to do this safely, we need to be certain that the encoding is
going to succeed. The reason this was previously encoded on the stack
first is to make sure that the function didn't corrupt people's data by
terminating due to an error. This shouldn't be possible when we force
TValue to be an unsigned type (avoiding the sort of arithmetic
right-shifting bugs that plagued PM in the 0.16 days).